+++ models/feeProject/PeriodUsage.js
@@ -1,15 +1,15 @@
 module.exports = (sequelize, DataTypes) => (
     sequelize.define('PeriodUsage', { // 정기검사 수수료
         sum: { // 합계 
-            type: DataTypes.INTEGER(11),
+            type: DataTypes.BIGINT,
             allowNull: true,
         },
         vat: { // 부가세
-            type: DataTypes.INTEGER(11),
+            type: DataTypes.BIGINT,
             allowNull: true,
         },
         totalCost: { // 총 비용
-            type: DataTypes.INTEGER(11),
+            type: DataTypes.BIGINT,
             allowNull: true,
         },
         offCheck: { // 야간, 공휴일, 토요일 검사시
